+++ src/main/kotlin/com/sipos/mmtrackerbackend/controller/GameController.kt
@@ -1,7 +1,12 @@
 package com.sipos.mmtrackerbackend.controller
 
+import com.sipos.mmtrackerbackend.dto.GameConverter
 import com.sipos.mmtrackerbackend.dto.GameDTORequest
-import com.sipos.mmtrackerbackend.service.GameService
+import com.sipos.mmtrackerbackend.dto.GameDTOResponse
+import com.sipos.mmtrackerbackend.model.Game
+import com.sipos.mmtrackerbackend.repository.GameRepository
+import com.sipos.mmtrackerbackend.repository.MapRepository
+import com.sipos.mmtrackerbackend.repository.UserRepository
 import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -9,7 +14,12 @@ import javax.persistence.EntityNotFoundException
 
 @RestController
 @RequestMapping("/api/games")
-class GameController(private val gameService: GameService) {
+class GameController(
+    private val gameRepository: GameRepository,
+    private val gameConverter: GameConverter,
+    private val userRepository: UserRepository,
+    private val mapRepository: MapRepository
+) {
 
     @ExceptionHandler(EntityNotFoundException::class, EmptyResultDataAccessException::class)
     fun handleException(): ResponseEntity<Unit> {
@@ -17,17 +27,40 @@ class GameController(private val gameService: GameService) {
     }
 
     @GetMapping
-    fun findAll() = gameService.findAll()
-
-    @PostMapping
-    fun addGame(@RequestBody game: GameDTORequest) = gameService.add(game)
+    fun findAll(): ResponseEntity<List<GameDTOResponse>> {
+        val games = gameRepository.findAll()
+        val response = mutableListOf<GameDTOResponse>()
+        games.forEach { game: Game ->
+            response.add(gameConverter.convertToResponse(game))
+        }
+        return ResponseEntity.ok(response)
+    }
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable id: Long) = gameService.getById(id)
+    fun getById(@PathVariable id: Long): ResponseEntity<GameDTOResponse> {
+        val gameToBeReturned = gameRepository.getOne(id)
+        return ResponseEntity.ok(gameConverter.convertToResponse(gameToBeReturned))
+    }
 
     @PutMapping("/{id}")
-    fun updateById(@PathVariable id: Long, @RequestBody game: GameDTORequest) = gameService.updateById(game, id)
+    fun updateById(@PathVariable id: Long, @RequestBody game: GameDTORequest): ResponseEntity<GameDTOResponse> {
+        val user = userRepository.getOne(game.user_id)
+        val map = mapRepository.getOne(game.map_id)
+        val gameToBeModified = gameRepository.getOne(id)
+        gameToBeModified.roundsWon = game.roundsWon
+        gameToBeModified.roundsLost = game.roundsLost
+        gameToBeModified.kills = game.kills
+        gameToBeModified.assists = game.assists
+        gameToBeModified.deaths = game.deaths
+        gameToBeModified.user = user
+        gameToBeModified.map = map
+        val savedGame = gameRepository.save(gameToBeModified)
+        return ResponseEntity.ok(gameConverter.convertToResponse(savedGame))
+    }
 
     @DeleteMapping("/{id}")
-    fun deleteById(@PathVariable id: Long) = gameService.deleteById(id)
+    fun deleteById(@PathVariable id: Long): ResponseEntity<Unit> {
+        gameRepository.deleteById(id)
+        return ResponseEntity.ok().build()
+    }
 }
