+++ src/main/kotlin/com/sipos/mmtrackerbackend/controller/MapController.kt
@@ -1,7 +1,10 @@
 package com.sipos.mmtrackerbackend.controller
 
+import com.sipos.mmtrackerbackend.dto.MapConverter
 import com.sipos.mmtrackerbackend.dto.MapDTORequest
-import com.sipos.mmtrackerbackend.service.MapService
+import com.sipos.mmtrackerbackend.dto.MapDTOResponse
+import com.sipos.mmtrackerbackend.model.Map
+import com.sipos.mmtrackerbackend.repository.MapRepository
 import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -9,7 +12,10 @@ import javax.persistence.EntityNotFoundException
 
 @RestController
 @RequestMapping("/api/maps")
-class MapController(private val mapService: MapService) {
+class MapController(
+    private val mapRepository: MapRepository,
+    private val mapConverter: MapConverter
+) {
 
     @ExceptionHandler(EntityNotFoundException::class, EmptyResultDataAccessException::class)
     fun handleException(): ResponseEntity<Unit> {
@@ -17,17 +23,47 @@ class MapController(private val mapService: MapService) {
     }
 
     @GetMapping
-    fun findAll() = mapService.findAll()
+    fun findAll(): ResponseEntity<List<MapDTOResponse>> {
+        val maps = mapRepository.findAll()
+        val response = mutableListOf<MapDTOResponse>()
+        maps.forEach { map: Map ->
+            response.add(mapConverter.convertToResponse(map))
+        }
+        return ResponseEntity.ok(response)
+    }
 
     @PostMapping
-    fun addMap(@RequestBody map: MapDTORequest) = mapService.add(map)
+    fun addMap(@RequestBody map: MapDTORequest): ResponseEntity<MapDTOResponse> {
+        if (mapRepository.existsByName(map.name)) {
+            val response = mapConverter.convertToResponse(mapRepository.findByName(map.name))
+            return ResponseEntity.ok(response)
+        }
+        if (mapRepository.existsByFileName(map.fileName)) {
+            val response = mapConverter.convertToResponse(mapRepository.findByFileName(map.fileName))
+            return ResponseEntity.ok(response)
+        }
+        val savedMap = mapRepository.save(Map(fileName = map.fileName, name = map.name))
+        return ResponseEntity.ok(mapConverter.convertToResponse(savedMap))
+    }
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable id: Long) = mapService.getById(id)
+    fun getById(@PathVariable id: Long): ResponseEntity<MapDTOResponse> {
+        val mapToBeReturned = mapRepository.getOne(id)
+        return ResponseEntity.ok(mapConverter.convertToResponse(mapToBeReturned))
+    }
 
     @PutMapping("/{id}")
-    fun updateById(@PathVariable id: Long, @RequestBody map: MapDTORequest) = mapService.updateById(map, id)
+    fun updateById(@PathVariable id: Long, @RequestBody map: MapDTORequest): ResponseEntity<MapDTOResponse> {
+        val mapToBeUpdated = mapRepository.getOne(id)
+        mapToBeUpdated.fileName = map.fileName
+        mapToBeUpdated.name = map.name
+        val savedMap = mapRepository.save(mapToBeUpdated)
+        return ResponseEntity.ok(mapConverter.convertToResponse(savedMap))
+    }
 
     @DeleteMapping("/{id}")
-    fun deleteById(@PathVariable id: Long) = mapService.deleteById(id)
+    fun deleteById(@PathVariable id: Long): ResponseEntity<Unit> {
+        mapRepository.deleteById(id)
+        return ResponseEntity.ok().build()
+    }
 }
