+++ src/main/kotlin/com/sipos/mmtrackerbackend/controller/UserController.kt
@@ -1,8 +1,11 @@
 package com.sipos.mmtrackerbackend.controller
 
-import com.sipos.mmtrackerbackend.dto.GameDTORequest
-import com.sipos.mmtrackerbackend.service.GameService
-import com.sipos.mmtrackerbackend.service.UserService
+import com.sipos.mmtrackerbackend.dto.*
+import com.sipos.mmtrackerbackend.model.Game
+import com.sipos.mmtrackerbackend.model.User
+import com.sipos.mmtrackerbackend.repository.GameRepository
+import com.sipos.mmtrackerbackend.repository.MapRepository
+import com.sipos.mmtrackerbackend.repository.UserRepository
 import org.springframework.dao.EmptyResultDataAccessException
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
@@ -10,7 +13,13 @@ import javax.persistence.EntityNotFoundException
 
 @RestController
 @RequestMapping("/api/users")
-class UserController(private val userService: UserService, val gameService: GameService) {
+class UserController(
+    private val userRepository: UserRepository,
+    private val userConverter: UserConverter,
+    private val gameRepository: GameRepository,
+    private val gameConverter: GameConverter,
+    private val mapRepository: MapRepository
+) {
 
     @ExceptionHandler(EntityNotFoundException::class, EmptyResultDataAccessException::class)
     fun handleException(): ResponseEntity<Unit> {
@@ -18,17 +27,52 @@ class UserController(private val userService: UserService, val gameService: Game
     }
 
     @GetMapping
-    fun findAll() = userService.findAll()
+    fun findAll(): ResponseEntity<List<UserDTOResponse>> {
+        val maps = userRepository.findAll()
+        val response = mutableListOf<UserDTOResponse>()
+        maps.forEach { user: User ->
+            response.add(userConverter.convertToResponse(user))
+        }
+        return ResponseEntity.ok(response)
+    }
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable id: Long) = userService.getById(id)
+    fun getById(@PathVariable id: Long): ResponseEntity<UserDTOResponse> {
+        val userToBeReturned = userRepository.getOne(id)
+        return ResponseEntity.ok(userConverter.convertToResponse(userToBeReturned))
+    }
 
     @DeleteMapping("/{id}")
-    fun deleteById(@PathVariable id: Long) = userService.deleteById(id)
+    fun deleteById(@PathVariable id: Long): ResponseEntity<Unit> {
+        userRepository.deleteById(id)
+        return ResponseEntity.ok().build()
+    }
 
     @GetMapping("/{id}/games")
-    fun getGames(@PathVariable id: Long) = gameService.getGamesOfUser(id)
+    fun getGames(@PathVariable id: Long): ResponseEntity<List<GameDTOResponse>> {
+        val games = gameRepository.findByUserId(id)
+        val response = mutableListOf<GameDTOResponse>()
+        games.forEach { game: Game ->
+            response.add(gameConverter.convertToResponse(game))
+        }
+        return ResponseEntity.ok(response)
+    }
 
     @PostMapping("/{id}/games")
-    fun addGame(@PathVariable id: Long, @RequestBody game: GameDTORequest) = gameService.addGameToUser(id, game)
+    fun addGame(@PathVariable id: Long, @RequestBody game: GameDTORequest): ResponseEntity<GameDTOResponse> {
+        val user = userRepository.getOne(id)
+        val map = mapRepository.getOne(game.map_id)
+        val gameSaved = gameRepository.save(
+            Game(
+                roundsWon = game.roundsWon,
+                roundsLost = game.roundsLost,
+                kills = game.kills,
+                assists = game.assists,
+                deaths = game.deaths,
+                user = user,
+                map = map,
+            )
+        )
+        return ResponseEntity.ok(gameConverter.convertToResponse(gameSaved))
+    }
 }
