--- src/main/kotlin/com/sipos/mmtrackerbackend/dto/AccountDTO.kt
@@ -1,6 +1,0 @@
-package com.sipos.mmtrackerbackend.dto
-
-data class AccountDTO(
-    var username: String,
-    var password: String
-)
