+++ src/main/kotlin/com/sipos/mmtrackerbackend/model/User.kt
@@ -1,5 +1,6 @@
 package com.sipos.mmtrackerbackend.model
 
+import com.fasterxml.jackson.annotation.JsonIgnore
 import javax.persistence.*
 
 @Entity
@@ -8,6 +9,11 @@ data class User(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long? = null,
+    @Column(unique = true)
     var username: String,
-    var password: String
+    var password: String,
+
+    @JsonIgnore
+    @OneToMany(mappedBy = "user", cascade = [CascadeType.REMOVE])
+    var games: List<Game> = emptyList()
 )
