+++ src/main/kotlin/com/sipos/mmtrackerbackend/repository/MapRepository.kt
@@ -7,4 +7,8 @@ import org.springframework.stereotype.Repository
 @Repository
 interface MapRepository : JpaRepository<Map, Long> {
 
+    fun existsByName(mapName: String): Boolean
+    fun existsByFileName(fileName: String): Boolean
+    fun findByName(mapName: String): Map
+    fun findByFileName(fileName: String): Map
 }
